Core/Utilities: Move EventMap into its own source file. * Move implementation details to .cpp.
