+++ src/server/shared/PrecompiledHeaders/sharedPCH.h
@@ -7,3 +7,4 @@
 #include "Errors.h"
 #include "TypeList.h"
 #include "TaskScheduler.h"
+#include "EventMap.h"
